CSC applied a bit/`rand.h` removal/new starry background with parallax/main refactor-reorganization-split up: - chore: Added floating part to avoid floating point precision lost warning. - feat: Added parallax starry background - style: Applied a bit more of CSC - chore: Removed some white spaces. - chore: Removed unnecessary `rand.h`. - refactor: Reorganized and split main into distinct file better code management. - refactor: Split multiple function doing too much at once. For instance functions doing draw and update at the same time, or the function that was drawing the tree, updating adn drawing the grass and the drawing platform.
