+++ lobby/include/common.h
@@ -4,7 +4,6 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
-#include <rand.h>
 #include <time.h>
 
 #include "raylib.h"
