+++ lobby/include/config.h
@@ -1,13 +1,8 @@
 #ifndef CONFIG_H
 #define CONFIG_H
 
-#define WINDOW_WIDTH 800
-#define WINDOW_HEIGHT 600
-
-#define SKY_WIDTH 8000.0f
-#define SKY_HEIGHT 2000.0f
-
-#define X_LIMIT 1200.0f
+#define WINDOW_WIDTH 800.0f
+#define WINDOW_HEIGHT 600.0f
 
 #define SOUNDS_PATH "assets/sounds/"
