+++ lobby/include/constant.h
@@ -1,6 +1,17 @@
 #ifndef CONSTANT_H
 #define CONSTANT_H
 
+#define SKY_WIDTH 8000.0f
+#define SKY_HEIGHT 2000.0f
+
+#define X_LIMIT 1200.0f
+
+#define GROUND_Y      500.0f
+#define PLAT_H        20.0f
+#define STEP_Y        100.0f
+#define WALL_W        100.0f
+#define DOOR_H        200.0f
+
 #define GRAVITY 1200.0f
 #define MOVE_SPEED 300.0f
 #define JUMP_FORCE 500.0f
@@ -10,5 +21,8 @@
 #define MAX_JUMPS 2
 
 #define FRICTION 2000
+
+#define MAX_GRASS_BLADES 80000
+
 
 #endif
