+++ lobby/include/skin.h
@@ -1,0 +1,12 @@
+#ifndef SKIN_H
+#define SKIN_H
+
+#include "types.h"
+
+void drawMenuTextures(void);
+void choosePlayerTexture(Player_st* player);
+
+void drawSkinButton(void);
+void toggleSkinMenu(void);
+
+#endif // SKIN_H
